Revert "Merge pull request #2593 from metriport/develop"
This reverts commit 56c5602c096418b1d514b55dddc0cfaefe3ffd4f, reversing
changes made to 68d756c0c5b985e49bde1ce9e62b99895e7b1729.
