+++ packages/api/package.json
@@ -27,7 +27,8 @@
   "license": "ISC",
   "dependencies": {
     "@aws-sdk/client-cloudwatch": "^3.338.0",
-    "@medplum/core": "^2.2.10",
+    "@medplum/core": "^2.0.32",
+    "@medplum/fhirtypes": "^2.0.32",
     "@metriport/api-sdk": "file:packages/api-sdk",
     "@metriport/carequality-sdk": "file:packages/carequality-sdk",
     "@metriport/commonwell-sdk": "file:packages/commonwell-sdk",
@@ -72,7 +73,6 @@
   },
   "devDependencies": {
     "@faker-js/faker": "^8.0.2",
-    "@medplum/fhirtypes": "^2.2.10",
     "@ngrok/ngrok": "^1.2.0",
     "@tsconfig/node18": "^1.0.1",
     "@types/convert-units": "^2.3.5",
