+++ packages/api/src/command/medical/organization/__tests__/create-organization.test.ts
@@ -1,17 +1,17 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
 import { faker } from "@faker-js/faker";
 import { v4 as uuidv4 } from "uuid";
-import * as address from "../../../../domain/medical/address";
 import { makeOrganization } from "../../../../domain/medical/__tests__/organization";
-import * as cqCommands from "../../../../external/carequality";
-import * as cwCommands from "../../../../external/commonwell";
+import * as address from "../../../../domain/medical/address";
 import * as createTenant from "../../../../external/fhir/admin";
-import * as upsertOrgToFHIRServer from "../../../../external/fhir/organization/upsert-organization";
 import { OrganizationModel } from "../../../../models/medical/organization";
 import { makeOrganizationOID } from "../../../../shared/oid";
 import * as createId from "../../customer-sequence/create-id";
 import { createOrganization } from "../create-organization";
 import { addressWithCoordinates } from "./register-organization";
+import * as upsertOrgToFHIRServer from "../../../../external/fhir/organization/upsert-organization";
+import * as cwCommands from "../../../../external/commonwell";
+import * as cqCommands from "../../../../external/carequality";
 
 let createOrganizationId_mock: jest.SpyInstance;
 let createTenantIfNotExistsMock: jest.SpyInstance;
