+++ packages/api/src/command/medical/patient/__tests__/consolidated-get.test.ts
@@ -1,16 +1,26 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import { BundleEntry, Resource } from "@medplum/fhirtypes";
+import { BundleEntry, Observation, Resource } from "@medplum/fhirtypes";
+import { HapiFhirClient } from "@metriport/core/external/fhir/api/api-hapi";
+import * as fhirBundle from "@metriport/core/external/fhir/shared/bundle";
+import * as fhirReferences from "../../../../external/fhir/references/get-references";
 import { resourcesSearchableByPatient } from "@metriport/api-sdk";
 import { ISO_DATE } from "@metriport/shared/common/date";
 import dayjs from "dayjs";
 import { v4 as uuidv4 } from "uuid";
+import { makePatient } from "../../../../domain/medical/__tests__/patient";
 import {
-  filterOutPrelimDocRefs,
+  getConsolidatedPatientData,
   getCurrentConsolidatedProgress,
   getIsSameResources,
+  filterOutPrelimDocRefs,
 } from "../consolidated-get";
 import { makeConsolidatedQueryProgress, requestId } from "./store-query-cmd";
+import * as getPatient from "../get-patient";
 
+let getPatientOrFailMock: jest.SpyInstance;
+let fhir_searchResourcePages: jest.SpyInstance;
+let fhir_getReferencesFromResources: jest.SpyInstance;
+let fhir_getReferencesFromFHIR: jest.SpyInstance;
 const defaultConsolidatedProgress = makeConsolidatedQueryProgress({
   requestId: requestId,
   status: "processing",
@@ -25,235 +35,470 @@ jest.mock("../../../../models/medical/patient");
 
 beforeEach(() => {
   jest.restoreAllMocks();
+  getPatientOrFailMock = jest.spyOn(getPatient, "getPatientOrFail");
+  fhir_searchResourcePages = jest.spyOn(HapiFhirClient.prototype, "searchResourcePages");
+  fhir_getReferencesFromResources = jest.spyOn(fhirBundle, "getReferencesFromResources");
+  fhir_getReferencesFromFHIR = jest.spyOn(fhirReferences, "getReferencesFromFHIR");
 });
 
-describe("consolidated-get", () => {
-  describe("getCurrentConsolidatedProgress", () => {
-    it("return undefined when there is no consolidated progress", () => {
-      const resp = getCurrentConsolidatedProgress(undefined, {});
+describe("getConsolidatedPatientData", () => {
+  const cxId = uuidv4();
+  const patientId = uuidv4();
 
-      expect(resp).toBeUndefined();
+  it("paginates appropriately", async () => {
+    getPatientOrFailMock.mockReturnValueOnce(makePatient({ id: patientId }));
+
+    const returnedResource: Observation = {
+      resourceType: "Observation",
+      id: "1",
+      status: "final",
+      code: {
+        coding: [
+          {
+            system: "http://loinc.org",
+            code: "12345",
+            display: "Test",
+          },
+        ],
+      },
+    };
+    const generator = (async function* generateResources() {
+      yield [returnedResource];
+      yield [returnedResource];
+    })();
+    fhir_searchResourcePages.mockReturnValue(generator);
+
+    const resp = await getConsolidatedPatientData({
+      patient: { cxId, id: patientId },
+      resources: ["Observation"],
     });
 
-    it("return undefined when there is no consolidated progress and we have params", () => {
-      const resp = getCurrentConsolidatedProgress(undefined, {
+    expect(resp).toBeTruthy();
+    expect(resp.resourceType).toEqual(`Bundle`);
+    expect(resp.total).toEqual(2);
+    expect(resp.entry).toEqual(expect.arrayContaining([{ resource: returnedResource }]));
+    expect(fhir_searchResourcePages).toHaveBeenCalledTimes(1);
+  });
+
+  it("hydration full 3 levels", async () => {
+    getPatientOrFailMock.mockReturnValueOnce(makePatient({ id: patientId }));
+
+    const initialResource: Observation = {
+      resourceType: "Observation",
+      id: "1",
+      status: "final",
+      code: {
+        coding: [
+          {
+            system: "http://loinc.org",
+            code: "12345",
+            display: "Test",
+          },
+        ],
+      },
+      subject: { reference: "Observation/2" },
+    };
+    const generator = (async function* generateResources() {
+      yield [initialResource];
+    })();
+    fhir_searchResourcePages.mockReturnValue(generator);
+
+    const missingResourceBaseL1: Observation = {
+      resourceType: "Observation",
+      status: "final",
+      code: {
+        coding: [
+          {
+            system: "http://loinc.org",
+            code: "12345",
+            display: "Test",
+          },
+        ],
+      },
+      id: "2",
+      subject: { reference: "Observation/3" },
+    };
+    const missingL2: Observation = {
+      ...missingResourceBaseL1,
+      id: "3",
+      subject: { reference: "Observation/4" },
+    };
+    const missingL3: Observation = {
+      ...missingResourceBaseL1,
+      id: "4",
+      subject: { reference: "Observation/5" },
+    };
+    const missingL4: Observation = {
+      ...missingResourceBaseL1,
+      id: "5",
+      subject: { reference: "Observation/6" },
+    };
+    fhir_getReferencesFromFHIR
+      .mockReturnValueOnce([missingResourceBaseL1])
+      .mockReturnValueOnce([missingL2])
+      .mockReturnValueOnce([missingL3])
+      .mockReturnValueOnce([missingL4]);
+
+    const resp = await getConsolidatedPatientData({
+      patient: { cxId, id: patientId },
+      resources: ["Observation"],
+    });
+
+    expect(resp).toBeTruthy();
+    expect(resp.resourceType).toEqual(`Bundle`);
+    expect(resp.total).toEqual(4);
+    expect(resp.entry).toEqual(
+      expect.arrayContaining([
+        { resource: initialResource },
+        { resource: missingResourceBaseL1 },
+        { resource: missingL2 },
+        { resource: missingL3 },
+      ])
+    );
+    expect(fhir_searchResourcePages).toHaveBeenCalledTimes(1);
+    expect(fhir_getReferencesFromResources).toHaveBeenCalledTimes(3);
+    expect(fhir_getReferencesFromFHIR).toHaveBeenCalledTimes(3);
+  });
+  it("hydration full 2 levels", async () => {
+    getPatientOrFailMock.mockReturnValueOnce(makePatient({ id: patientId }));
+
+    const initialResource: Observation = {
+      resourceType: "Observation",
+      id: "1",
+      status: "final",
+      code: {
+        coding: [
+          {
+            system: "http://loinc.org",
+            code: "12345",
+            display: "Test",
+          },
+        ],
+      },
+      subject: { reference: "Observation/2" },
+    };
+    const generator = (async function* generateResources() {
+      yield [initialResource];
+    })();
+    fhir_searchResourcePages.mockReturnValue(generator);
+
+    const missingResourceBaseL1: Observation = {
+      resourceType: "Observation",
+      status: "final",
+      code: {
+        coding: [
+          {
+            system: "http://loinc.org",
+            code: "12345",
+            display: "Test",
+          },
+        ],
+      },
+      id: "2",
+      subject: { reference: "Observation/3" },
+    };
+    const missingL2: Observation = {
+      ...missingResourceBaseL1,
+      id: "3",
+    };
+    const missingL3: Observation = {
+      ...missingResourceBaseL1,
+      id: "4",
+      subject: { reference: "Observation/5" },
+    };
+    fhir_getReferencesFromFHIR
+      .mockReturnValueOnce([missingResourceBaseL1])
+      .mockReturnValueOnce([missingL2])
+      .mockReturnValueOnce([missingL3]);
+
+    const resp = await getConsolidatedPatientData({
+      patient: { cxId, id: patientId },
+      resources: ["Observation"],
+    });
+
+    expect(resp).toBeTruthy();
+    expect(resp.resourceType).toEqual(`Bundle`);
+    expect(resp.total).toEqual(3);
+    expect(resp.entry).toEqual(
+      expect.arrayContaining([
+        { resource: initialResource },
+        { resource: missingResourceBaseL1 },
+        { resource: missingL2 },
+      ])
+    );
+    expect(fhir_searchResourcePages).toHaveBeenCalledTimes(1);
+    expect(fhir_getReferencesFromResources).toHaveBeenCalledTimes(3);
+    expect(fhir_getReferencesFromFHIR).toHaveBeenCalledTimes(2);
+  });
+});
+
+describe("getCurrentConsolidatedProgress", () => {
+  it("return undefined when there is no consolidated progress", () => {
+    const resp = getCurrentConsolidatedProgress(undefined, {});
+
+    expect(resp).toBeUndefined();
+  });
+
+  it("return undefined when there is no consolidated progress and we have params", () => {
+    const resp = getCurrentConsolidatedProgress(undefined, {
+      resources: ["Observation"],
+      dateFrom: dayjs().subtract(10, "years").format(ISO_DATE),
+      dateTo: dayjs().add(1, "day").format(ISO_DATE),
+      conversionType: "json",
+    });
+
+    expect(resp).toBeUndefined();
+  });
+
+  it("return undefined when consolidated progress empty array and we have params", () => {
+    const resp = getCurrentConsolidatedProgress([], {
+      resources: ["Observation"],
+      dateFrom: dayjs().subtract(10, "years").format(ISO_DATE),
+      dateTo: dayjs().add(1, "day").format(ISO_DATE),
+      conversionType: "json",
+    });
+
+    expect(resp).toBeUndefined();
+  });
+
+  it("return undefined when consolidated progress is present but params are empty", () => {
+    const consolidatedQueryProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
+    const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      requestId: uuidv4(),
+      status: "completed",
+      conversionType: "pdf",
+    });
+    const resp = getCurrentConsolidatedProgress(
+      [consolidatedQueryProgress, secondConsolidatedQueryProgress],
+      {}
+    );
+
+    expect(resp).toBeUndefined();
+  });
+
+  it("return undefined when consolidated progress is present but params dont match", () => {
+    const consolidatedQueryProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
+    const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      requestId: uuidv4(),
+      status: "completed",
+      conversionType: "pdf",
+    });
+    const resp = getCurrentConsolidatedProgress(
+      [consolidatedQueryProgress, secondConsolidatedQueryProgress],
+      {
         resources: ["Observation"],
         dateFrom: dayjs().subtract(10, "years").format(ISO_DATE),
         dateTo: dayjs().add(1, "day").format(ISO_DATE),
-        conversionType: "json",
-      });
-
-      expect(resp).toBeUndefined();
-    });
-
-    it("return undefined when consolidated progress empty array and we have params", () => {
-      const resp = getCurrentConsolidatedProgress([], {
-        resources: ["Observation"],
-        dateFrom: dayjs().subtract(10, "years").format(ISO_DATE),
-        dateTo: dayjs().add(1, "day").format(ISO_DATE),
-        conversionType: "json",
-      });
-
-      expect(resp).toBeUndefined();
-    });
-
-    it("return undefined when consolidated progress is present but params are empty", () => {
-      const consolidatedQueryProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
-      const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        requestId: uuidv4(),
-        status: "completed",
         conversionType: "pdf",
-      });
-      const resp = getCurrentConsolidatedProgress(
-        [consolidatedQueryProgress, secondConsolidatedQueryProgress],
-        {}
-      );
+      }
+    );
 
-      expect(resp).toBeUndefined();
+    expect(resp).toBeUndefined();
+  });
+
+  it("return undefined when consolidated progress is present and params match but is completed", () => {
+    const consolidatedQueryProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      status: "completed",
     });
-
-    it("return undefined when consolidated progress is present but params dont match", () => {
-      const consolidatedQueryProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
-      const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        requestId: uuidv4(),
-        status: "completed",
-        conversionType: "pdf",
-      });
-      const resp = getCurrentConsolidatedProgress(
-        [consolidatedQueryProgress, secondConsolidatedQueryProgress],
-        {
-          resources: ["Observation"],
-          dateFrom: dayjs().subtract(10, "years").format(ISO_DATE),
-          dateTo: dayjs().add(1, "day").format(ISO_DATE),
-          conversionType: "pdf",
-        }
-      );
-
-      expect(resp).toBeUndefined();
+    const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      requestId: uuidv4(),
+      status: "completed",
+      conversionType: "pdf",
     });
+    const resp = getCurrentConsolidatedProgress(
+      [consolidatedQueryProgress, secondConsolidatedQueryProgress],
+      {
+        resources: consolidatedQueryProgress.resources,
+        dateFrom: consolidatedQueryProgress.dateFrom,
+        dateTo: consolidatedQueryProgress.dateTo,
+        conversionType: consolidatedQueryProgress.conversionType,
+      }
+    );
 
-    it("return undefined when consolidated progress is present and params match but is completed", () => {
-      const consolidatedQueryProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        status: "completed",
-      });
-      const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        requestId: uuidv4(),
-        status: "completed",
-        conversionType: "pdf",
-      });
-      const resp = getCurrentConsolidatedProgress(
-        [consolidatedQueryProgress, secondConsolidatedQueryProgress],
-        {
-          resources: consolidatedQueryProgress.resources,
-          dateFrom: consolidatedQueryProgress.dateFrom,
-          dateTo: consolidatedQueryProgress.dateTo,
-          conversionType: consolidatedQueryProgress.conversionType,
-        }
-      );
+    expect(resp).toBeUndefined();
+  });
 
-      expect(resp).toBeUndefined();
+  it("return consolidated progress when consolidated progress is present and params match", () => {
+    const consolidatedProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
+    const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      requestId: uuidv4(),
+      status: "completed",
+      conversionType: "pdf",
     });
+    const resp = getCurrentConsolidatedProgress(
+      [consolidatedProgress, secondConsolidatedQueryProgress],
+      {
+        resources: consolidatedProgress.resources,
+        dateFrom: consolidatedProgress.dateFrom,
+        dateTo: consolidatedProgress.dateTo,
+        conversionType: consolidatedProgress.conversionType,
+      }
+    );
 
-    it("return consolidated progress when consolidated progress is present and params match", () => {
-      const consolidatedProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
-      const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        requestId: uuidv4(),
-        status: "completed",
-        conversionType: "pdf",
-      });
-      const resp = getCurrentConsolidatedProgress(
-        [consolidatedProgress, secondConsolidatedQueryProgress],
-        {
-          resources: consolidatedProgress.resources,
-          dateFrom: consolidatedProgress.dateFrom,
-          dateTo: consolidatedProgress.dateTo,
-          conversionType: consolidatedProgress.conversionType,
-        }
-      );
+    expect(resp).toEqual(consolidatedProgress);
+  });
 
-      expect(resp).toEqual(consolidatedProgress);
+  it("return consolidated progress when consolidated progress is present with empty resource array and resources is the full list", () => {
+    const consolidatedProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      resources: [...resourcesSearchableByPatient],
     });
-
-    it("return consolidated progress when consolidated progress is present with empty resource array and resources is the full list", () => {
-      const consolidatedProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        resources: [...resourcesSearchableByPatient],
-      });
-      const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        requestId: uuidv4(),
-        resources: ["Observation"],
-      });
-      const resp = getCurrentConsolidatedProgress(
-        [consolidatedProgress, secondConsolidatedQueryProgress],
-        {
-          resources: [],
-          dateFrom: consolidatedProgress.dateFrom,
-          dateTo: consolidatedProgress.dateTo,
-          conversionType: consolidatedProgress.conversionType,
-        }
-      );
-
-      expect(resp).toEqual(consolidatedProgress);
+    const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      requestId: uuidv4(),
+      resources: ["Observation"],
     });
-
-    it("return consolidated progress when consolidated progress is present with full resource array and resources is empty", () => {
-      const consolidatedProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
-      const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        requestId: uuidv4(),
-        resources: ["Observation"],
-      });
-      const resp = getCurrentConsolidatedProgress(
-        [consolidatedProgress, secondConsolidatedQueryProgress],
-        {
-          resources: [...resourcesSearchableByPatient],
-          dateFrom: consolidatedProgress.dateFrom,
-          dateTo: consolidatedProgress.dateTo,
-          conversionType: consolidatedProgress.conversionType,
-        }
-      );
-
-      expect(resp).toEqual(consolidatedProgress);
-    });
-
-    it("return consolidatedProgressTwo when params are equal to consolidatedProgressTwo", () => {
-      const consolidatedProgressOne = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        resources: ["Observation"],
-        conversionType: "json",
-      });
-      const consolidatedProgressTwo = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
+    const resp = getCurrentConsolidatedProgress(
+      [consolidatedProgress, secondConsolidatedQueryProgress],
+      {
         resources: [],
-        conversionType: "pdf",
-      });
-      const consolidatedProgressThree = makeConsolidatedQueryProgress({
-        ...defaultConsolidatedProgress,
-        resources: ["Observation", "AllergyIntolerance", "Practitioner"],
-        conversionType: "html",
-      });
+        dateFrom: consolidatedProgress.dateFrom,
+        dateTo: consolidatedProgress.dateTo,
+        conversionType: consolidatedProgress.conversionType,
+      }
+    );
 
-      const resp = getCurrentConsolidatedProgress(
-        [consolidatedProgressOne, consolidatedProgressTwo, consolidatedProgressThree],
+    expect(resp).toEqual(consolidatedProgress);
+  });
+
+  it("return consolidated progress when consolidated progress is present with full resource array and resources is empty", () => {
+    const consolidatedProgress = makeConsolidatedQueryProgress(defaultConsolidatedProgress);
+    const secondConsolidatedQueryProgress = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      requestId: uuidv4(),
+      resources: ["Observation"],
+    });
+    const resp = getCurrentConsolidatedProgress(
+      [consolidatedProgress, secondConsolidatedQueryProgress],
+      {
+        resources: [...resourcesSearchableByPatient],
+        dateFrom: consolidatedProgress.dateFrom,
+        dateTo: consolidatedProgress.dateTo,
+        conversionType: consolidatedProgress.conversionType,
+      }
+    );
+
+    expect(resp).toEqual(consolidatedProgress);
+  });
+
+  it("return consolidatedProgressTwo when params are equal to consolidatedProgressTwo", () => {
+    const consolidatedProgressOne = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      resources: ["Observation"],
+      conversionType: "json",
+    });
+    const consolidatedProgressTwo = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      resources: [],
+      conversionType: "pdf",
+    });
+    const consolidatedProgressThree = makeConsolidatedQueryProgress({
+      ...defaultConsolidatedProgress,
+      resources: ["Observation", "AllergyIntolerance", "Practitioner"],
+      conversionType: "html",
+    });
+
+    const resp = getCurrentConsolidatedProgress(
+      [consolidatedProgressOne, consolidatedProgressTwo, consolidatedProgressThree],
+      {
+        resources: consolidatedProgressTwo.resources,
+        dateFrom: consolidatedProgressTwo.dateFrom,
+        dateTo: consolidatedProgressTwo.dateTo,
+        conversionType: consolidatedProgressTwo.conversionType,
+      }
+    );
+
+    expect(resp).toEqual(consolidatedProgressTwo);
+  });
+});
+
+describe("getIsSameResources", () => {
+  it("return true when both arrays are empty", () => {
+    const resp = getIsSameResources([], []);
+
+    expect(resp).toBeTruthy();
+  });
+
+  it("return false when one of the arrays is empty", () => {
+    const resp = getIsSameResources([], ["Observation"]);
+
+    expect(resp).toBeFalsy();
+  });
+
+  it("return false when both arrays are not equal", () => {
+    const resp = getIsSameResources(["Observation"], ["AllergyIntolerance"]);
+
+    expect(resp).toBeFalsy();
+  });
+
+  it("return true when both arrays are equal", () => {
+    const resp = getIsSameResources(
+      ["Observation", "AllergyIntolerance"],
+      ["AllergyIntolerance", "Observation"]
+    );
+
+    expect(resp).toBeTruthy();
+  });
+
+  it("return true when first array is empty and the other is resourcesSearchableByPatient", () => {
+    const resp = getIsSameResources([], [...resourcesSearchableByPatient]);
+
+    expect(resp).toBeTruthy();
+  });
+
+  it("return true when second array is empty and the other is resourcesSearchableByPatient", () => {
+    const resp = getIsSameResources([...resourcesSearchableByPatient], []);
+
+    expect(resp).toBeTruthy();
+  });
+});
+
+describe("filterOutPrelimDocRefs", () => {
+  const prelimDocRef: BundleEntry<Resource> = {
+    resource: {
+      resourceType: "DocumentReference",
+      id: "1",
+      docStatus: "preliminary",
+      content: [
         {
-          resources: consolidatedProgressTwo.resources,
-          dateFrom: consolidatedProgressTwo.dateFrom,
-          dateTo: consolidatedProgressTwo.dateTo,
-          conversionType: consolidatedProgressTwo.conversionType,
-        }
-      );
+          attachment: {
+            contentType: "application/pdf",
+            url: "http://test.com",
+          },
+        },
+      ],
+    },
+  };
 
-      expect(resp).toEqual(consolidatedProgressTwo);
-    });
+  it("return empty array when resources is empty", () => {
+    const resp = filterOutPrelimDocRefs([]);
+
+    expect(resp).toEqual([]);
   });
 
-  describe("getIsSameResources", () => {
-    it("return true when both arrays are empty", () => {
-      const resp = getIsSameResources([], []);
-      expect(resp).toBeTruthy();
-    });
+  it("return undefined when resources is undefined", () => {
+    const resp = filterOutPrelimDocRefs(undefined);
 
-    it("return false when one of the arrays is empty", () => {
-      const resp = getIsSameResources([], ["Observation"]);
-      expect(resp).toBeFalsy();
-    });
-
-    it("return false when both arrays are not equal", () => {
-      const resp = getIsSameResources(["Observation"], ["AllergyIntolerance"]);
-      expect(resp).toBeFalsy();
-    });
-
-    it("return true when both arrays are equal", () => {
-      const resp = getIsSameResources(
-        ["Observation", "AllergyIntolerance"],
-        ["AllergyIntolerance", "Observation"]
-      );
-      expect(resp).toBeTruthy();
-    });
-
-    it("return true when first array is empty and the other is resourcesSearchableByPatient", () => {
-      const resp = getIsSameResources([], [...resourcesSearchableByPatient]);
-      expect(resp).toBeTruthy();
-    });
-
-    it("return true when second array is empty and the other is resourcesSearchableByPatient", () => {
-      const resp = getIsSameResources([...resourcesSearchableByPatient], []);
-      expect(resp).toBeTruthy();
-    });
+    expect(resp).toEqual(undefined);
   });
 
-  describe("filterOutPrelimDocRefs", () => {
-    const prelimDocRef: BundleEntry<Resource> = {
+  it("return empty array when resources has only prelim doc refs", () => {
+    const resp = filterOutPrelimDocRefs([prelimDocRef, { ...prelimDocRef, id: "2" }]);
+
+    expect(resp).toEqual([]);
+  });
+
+  it("return empty array when resources has prelim doc refs", () => {
+    const finalDocRef: BundleEntry<Resource> = {
       resource: {
         resourceType: "DocumentReference",
-        id: "1",
-        docStatus: "preliminary",
+        id: "2",
+        docStatus: "final",
         content: [
           {
             attachment: {
@@ -265,39 +510,8 @@ describe("consolidated-get", () => {
       },
     };
 
-    it("return empty array when resources is empty", () => {
-      const resp = filterOutPrelimDocRefs([]);
-      expect(resp).toEqual([]);
-    });
+    const resp = filterOutPrelimDocRefs([prelimDocRef, finalDocRef]);
 
-    it("return undefined when resources is undefined", () => {
-      const resp = filterOutPrelimDocRefs(undefined);
-      expect(resp).toEqual(undefined);
-    });
-
-    it("return empty array when resources has only prelim doc refs", () => {
-      const resp = filterOutPrelimDocRefs([prelimDocRef, { ...prelimDocRef, id: "2" }]);
-      expect(resp).toEqual([]);
-    });
-
-    it("return empty array when resources has prelim doc refs", () => {
-      const finalDocRef: BundleEntry<Resource> = {
-        resource: {
-          resourceType: "DocumentReference",
-          id: "2",
-          docStatus: "final",
-          content: [
-            {
-              attachment: {
-                contentType: "application/pdf",
-                url: "http://test.com",
-              },
-            },
-          ],
-        },
-      };
-      const resp = filterOutPrelimDocRefs([prelimDocRef, finalDocRef]);
-      expect(resp).toEqual([finalDocRef]);
-    });
+    expect(resp).toEqual([finalDocRef]);
   });
 });
