+++ packages/api/src/command/medical/patient/convert-fhir-bundle.ts
@@ -17,9 +17,8 @@ import { Config } from "../../../shared/config";
 import { getSandboxSeedData } from "../../../shared/sandbox/sandbox-seed-data";
 import { createSandboxMRSummaryFileName } from "./shared";
 
-dayjs.extend(duration);
-
 const s3Utils = new S3Utils(Config.getAWSRegion());
+dayjs.extend(duration);
 
 /**
  * Keep this a couple seconds higher than the respective lambda's timeout.
