+++ packages/api/src/errors/index.ts
@@ -1,5 +1,3 @@
-import { out } from "@metriport/core/util";
-import { Config } from "../shared/config";
 import { errorToString } from "../shared/log";
 import { capture } from "../shared/notifications";
 
@@ -7,11 +5,9 @@ export function getErrorMessage(error: unknown) {
   return errorToString(error);
 }
 
-export function processAsyncError(msg: string, useMsgAsTitle = false) {
-  const { log } = out("processAsyncError");
-  return (error: unknown) => {
-    if (Config.isDev()) log(`${msg}:`, error);
-    else log(`${msg}: ${getErrorMessage(error)}`);
-    capture.error(useMsgAsTitle ? msg : error, { extra: { message: msg, error } });
+export function processAsyncError(msg: string) {
+  return (err: unknown) => {
+    console.error(`${msg}: ${getErrorMessage(err)}`);
+    capture.error(err, { extra: { message: msg, err } });
   };
 }
