+++ packages/core/package.json
@@ -116,7 +116,7 @@
     "xmldom": "^0.6.0"
   },
   "devDependencies": {
-    "@medplum/fhirtypes": "^2.2.10",
+    "@medplum/fhirtypes": "^2.0.32",
     "@tsconfig/recommended": "^1.0.2",
     "@types/formidable": "^3.4.5",
     "@types/jest": "29.5.3",
