--- packages/core/src/command/consolidated/consolidated-connector-factory.ts
@@ -1,13 +1,0 @@
-import { Config } from "../../util/config";
-import { ConsolidatedDataConnector } from "./consolidated-connector";
-import { ConsolidatedDataConnectorLambda } from "./consolidated-connector-lambda";
-import { ConsolidatedDataConnectorLocal } from "./consolidated-connector-local";
-
-export function buildConsolidatedDataConnector(): ConsolidatedDataConnector {
-  if (Config.isDev()) {
-    const bucketName = Config.getMedicalDocumentsBucketName();
-    const apiURL = Config.getApiUrl();
-    return new ConsolidatedDataConnectorLocal(bucketName, apiURL);
-  }
-  return new ConsolidatedDataConnectorLambda();
-}
