+++ packages/core/src/util/config.ts
@@ -107,11 +107,6 @@ export class Config {
     return getEnvVar("IHE_PARSED_RESPONSES_BUCKET_NAME");
   }
 
-  // TODO 1319 Move this to required
-  static getFHIRtoBundleLambdaName(): string | undefined {
-    return getEnvVar("FHIR_TO_BUNDLE_LAMBDA_NAME");
-  }
-
   static getBedrockRegion(): string | undefined {
     return getEnvVar("BEDROCK_REGION");
   }
