+++ packages/lambdas/package.json
@@ -26,7 +26,8 @@
     "@aws-sdk/client-s3": "^3.435.0",
     "@aws-sdk/client-secrets-manager": "^3.348.0",
     "@aws-sdk/s3-request-presigner": "^3.435.0",
-    "@medplum/core": "^2.2.10",
+    "@medplum/core": "^2.0.32",
+    "@medplum/fhirtypes": "^2.0.32",
     "@metriport/api-sdk": "^7.12.7",
     "@metriport/commonwell-sdk": "^4.12.10",
     "@opensearch-project/opensearch": "^2.3.1",
@@ -60,7 +61,6 @@
     "zod": "^3.22.4"
   },
   "devDependencies": {
-    "@medplum/fhirtypes": "^2.2.10",
     "@metriport/core": "file:packages/core",
     "@metriport/ihe-gateway-sdk": "file:packages/ihe-gateway-sdk",
     "@metriport/shared": "file:packages/shared",
