+++ packages/shared/src/index.ts
@@ -20,7 +20,6 @@ export { BadRequestError } from "./error/bad-request";
 export { MetriportError } from "./error/metriport-error";
 export { NotFoundError } from "./error/not-found";
 export { errorToString } from "./error/shared";
-export * from "./interface";
 export * as medical from "./medical";
 export * from "./net/error";
 export * from "./net/retry";
