+++ packages/shared/src/interface/index.ts
@@ -1,2 +1,1 @@
-export * from "./internal";
 export * from "./schema";
