--- packages/shared/src/interface/internal/index.ts
@@ -1,1 +1,0 @@
-export * from "./consolidated";
