--- packages/shared/src/medical/fhir/conversion.ts
@@ -1,3 +1,0 @@
-export const consolidationConversionType = ["html", "pdf", "json"] as const;
-
-export type ConsolidationConversionType = (typeof consolidationConversionType)[number];
