--- packages/shared/src/medical/fhir/resources.ts
@@ -1,76 +1,0 @@
-export const resourcesSearchableByPatient = [
-  "Account",
-  "AllergyIntolerance",
-  "Appointment",
-  "AppointmentResponse",
-  "AuditEvent",
-  "Basic",
-  "BodyStructure",
-  "CarePlan",
-  "CareTeam",
-  "ChargeItem",
-  "Claim",
-  "ClaimResponse",
-  "ClinicalImpression",
-  "Communication",
-  "CommunicationRequest",
-  "Composition",
-  "Condition",
-  "Consent",
-  "Contract",
-  "Coverage",
-  "CoverageEligibilityRequest",
-  "CoverageEligibilityResponse",
-  "DetectedIssue",
-  "Device",
-  "DeviceRequest",
-  "DeviceUseStatement",
-  "DiagnosticReport",
-  "DocumentManifest",
-  "DocumentReference",
-  "Encounter",
-  "EnrollmentRequest",
-  "EpisodeOfCare",
-  "ExplanationOfBenefit",
-  "FamilyMemberHistory",
-  "Flag",
-  "Goal",
-  "GuidanceResponse",
-  "ImagingStudy",
-  "Immunization",
-  "ImmunizationEvaluation",
-  "ImmunizationRecommendation",
-  "Invoice",
-  "List",
-  "MeasureReport",
-  "Media",
-  "MedicationAdministration",
-  "MedicationDispense",
-  "MedicationRequest",
-  "MedicationStatement",
-  "MolecularSequence",
-  "NutritionOrder",
-  "Observation",
-  "Person",
-  "Procedure",
-  "Provenance",
-  "QuestionnaireResponse",
-  "RelatedPerson",
-  "RequestGroup",
-  "ResearchSubject",
-  "RiskAssessment",
-  "ServiceRequest",
-  "Specimen",
-] as const;
-
-export const resourcesSearchableBySubject = ["AdverseEvent", "Task"] as const;
-
-export const generalResources = ["Practitioner"] as const;
-
-export const resourceTypeForConsolidation = [
-  ...resourcesSearchableByPatient,
-  ...resourcesSearchableBySubject,
-  ...generalResources,
-] as const;
-
-export type ResourceTypeForConsolidation = (typeof resourceTypeForConsolidation)[number];
