+++ packages/shared/src/medical/index.ts
@@ -1,5 +1,3 @@
 export * from "./fhir/bundle";
-export * from "./fhir/resources";
-export * from "./fhir/conversion";
 export * from "./webhook/webhook-request";
 export * from "./webhook/webhook-status-response";
