+++ packages/utils/package.json
@@ -38,7 +38,8 @@
     "@aws-sdk/client-bedrock-agent-runtime": "^3.616.0",
     "@aws-sdk/client-bedrock-runtime": "^3.616.0",
     "@aws-sdk/client-cognito-identity-provider": "^3.301.0",
-    "@medplum/core": "^2.2.10",
+    "@medplum/core": "^2.0.32",
+    "@medplum/fhirtypes": "^2.0.32",
     "@metriport/api-sdk": "file:packages/api-sdk",
     "@metriport/core": "file:packages/core",
     "@metriport/shared": "file:packages/shared",
@@ -61,7 +62,6 @@
     "sequelize": "^6.31.0"
   },
   "devDependencies": {
-    "@medplum/fhirtypes": "^2.2.10",
     "@tsconfig/recommended": "^1.0.1",
     "@types/aws-serverless-express": "^3.3.9",
     "@types/express": "^4.17.14",
